[FEAT] 매니저 대회 정보 수정 페이지 구현 (#290)
* chore: 불필요 game 경로 제거

* feat: LeagueForm 컴포넌트 분리

* fix: 대회 정보 수정 페이지에서 Round 정보가 연동되지 않는 문제 해결
