--- apps/manager/app/game/[leagueId]/[gameId]/score/[teamId]/page.tsx
@@ -1,3 +1,0 @@
-export default function ScoreEdit() {
-  return <>미컴 축구생각</>;
-}
