+++ apps/manager/app/league/_components/LeagueForm/LeagueForm.css.ts
@@ -1,18 +1,11 @@
 import { rem, theme } from '@hcc/styles';
 import { style } from '@vanilla-extract/css';
 
-export const layout = style({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'space-between',
-
-  height: '100%',
-  paddingInline: theme.sizes.appInlinePadding,
-  paddingTop: theme.sizes.appInlinePadding,
-});
-
 export const form = style({
   ...theme.layouts.column,
+  paddingInline: theme.sizes.appInlinePadding,
+  paddingTop: theme.sizes.appInlinePadding,
+
   gap: rem(12),
   flex: 1,
 });
